fix: V7 迁移 friend_approval 列类型 TINYINT → INT，修复 Hibernate schema-validation 不匹配
Kotlin Int 映射为 Types#INTEGER，对应 MySQL INT；TINYINT 导致校验失败

+++ repository/src/main/resources/db/migration/V7__add_friend_approval.sql
@@ -4,4 +4,4 @@
 -- 1=AUTO_ACCEPT（自动通过）
 -- 2=AUTO_REJECT（自动拒绝）
 
-ALTER TABLE users ADD COLUMN friend_approval TINYINT NOT NULL DEFAULT 0 AFTER privacy_status;
+ALTER TABLE users ADD COLUMN friend_approval INT NOT NULL DEFAULT 0 AFTER privacy_status;
